refactor: 테스트 패키지 구조 변경
