+++ src/test/java/sample/cafekiosk/spring/domain/api/OrderControllerTest.java
@@ -10,8 +10,8 @@ import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 import org.springframework.test.web.servlet.result.MockMvcResultHandlers;
-import sample.cafekiosk.spring.api.controller.OrderController;
-import sample.cafekiosk.spring.domain.product.dto.request.OrderCreateRequest;
+import sample.cafekiosk.spring.api.controller.order.OrderController;
+import sample.cafekiosk.spring.api.controller.order.dto.request.OrderCreateRequest;
 import sample.cafekiosk.spring.api.service.order.OrderService;
 
 import java.util.List;
