+++ src/test/java/sample/cafekiosk/spring/domain/api/ProductControllerTest.java
@@ -10,11 +10,11 @@ import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 import org.springframework.test.web.servlet.result.MockMvcResultHandlers;
-import sample.cafekiosk.spring.api.controller.ProductController;
+import sample.cafekiosk.spring.api.controller.product.ProductController;
 import sample.cafekiosk.spring.domain.product.ProductType;
 import sample.cafekiosk.spring.domain.product.SellingStatus;
-import sample.cafekiosk.spring.domain.product.dto.ProductResponse;
-import sample.cafekiosk.spring.domain.product.dto.request.ProductCreateRequest;
+import sample.cafekiosk.spring.api.controller.product.dto.response.ProductResponse;
+import sample.cafekiosk.spring.api.controller.product.dto.request.ProductCreateRequest;
 import sample.cafekiosk.spring.api.service.product.ProductService;
 
 import java.util.List;
